Convert dump_buf_hex() to use oc_text_buf instead of sprintf
I seem to recall the OpenBSD build will complain loudly about the use of
"bad" functions like sprintf. And even though this particular code does
seem to be perfectly correct, they do have a point in the general case.

Just use buf_append() for this, since that's what it was designed for.

Signed-off-by: David Woodhouse <dwmw2@infradead.org>