Fix proxy username and password parsing.
We are supposed to take the *last* (unescaped) @ sign as the separation
between user:pass and hostname, not the first. So use strrchr() instead
of strchr().

Conversely, the first colon is the separation between user and pass so
strchr is still correct there.

Also actually support unescaping the resulting username and password.

Signed-off-by: David Woodhouse <dwmw2@infradead.org>