Silence Coverity warning in cancellable_connect()
It's whining that we are reading a single byte into an int variable.
Which would be wrong if we were doing anything with it, and is often
a sign of little-endian-itis. As it happens it's harmless here because
it's only an attempt to capture the correct error condition by reading
from a defunct socket. But fix it anyway. There was no need for the
mismatch.

Signed-off-by: David Woodhouse <dwmw2@infradead.org>