oncp: Tear down ESP if receiving new config fails
Spotted by Coverity as an 'unused value' because the return value from
oncp_receive_espkeys() was ignored.

Use oncp_esp_close() to do the shutdown, and fix it only to send the
'ESP off' message to the server if an 'ESP on' message has previously
been sent (i.e. if vpninfo->dtls_state >= DTLS_CONNECTING).

Signed-off-by: David Woodhouse <dwmw2@infradead.org>