Fix potential NULL dereference in GlobalProtect XML parsing
Spotted by Coverity.

Perhaps the intent in reusing xml_node here for the inner loop was that
by leaving it NULL we'd automatically break out of the outer loop too,
which is reasonable. Except it wouldn't have done that because it'd
have died on trying to move to xmlnode->next. Break out of the outer
loop explicitly, instead.

Signed-off-by: David Woodhouse <dwmw2@infradead.org>