Correct handling of empty responses to HTTP requests
If there's nothing to return, read_http_header() doesn't append anything
at all to hdrbuf, and it ends up with hdrbuf->data still NULL.

As long as *something* has once been added to it, there is a guarantee
that it's OK to use hdrbuf->data as a NUL-terminated string, as we do
when looking for HTTP/1.x responses.

Spotted when I just hit 'enter' when testing with openssl s_server.

Fixes: 093de80507 ("Rework HTTP header fetching to receive into oc_text_buf")
Signed-off-by: David Woodhouse <dwmw2@infradead.org>