Always send client cert
TLS servers may request a certificate from the client. This request includes a list of 0 or more acceptable issuer DNs. The client may use this list to determine which certificate to send. GnuTLS's default behavior is to not send a client certificate if there is no match. However, we generally always have a specific certificate specified, so we just want to send that regardless.

Originally submitted as PR on GitHub: https://github.com/dlenski/openconnect/pull/164
Signed-off-by: Daniel Lenski <dlenski@gmail.com>