Run Cisco CSD script as child, not daemonized grandchild
This allows us to capture a failure in the CSD script/binary much more
quickly, rather than spinning endlessly (see #108 for one of many examples
where this confuses users).

Tested with both “real” CSD trojan binaries and wrapper script, as well as
`trojans/csd-post.sh`.

GP already does this for the HIP script, and it works fine, including with
NM and Android clients based on libopenconnect.

Signed-off-by: Daniel Lenski <dlenski@gmail.com>