ignore failure in downloading CSD stub if CSD wrapper is specified
Resolves the issue reported here:
	https://lists.infradead.org/pipermail/openconnect-devel/2020-March/005554.html

Basically, what's happened here is that the Cisco VPN admins have
misconfigured things so that they require running CSD on all platforms, but
the CSD “stub” script specified for Linux _doesn't actually exist_.  (They
probably only tested with Mac, Windows, and Android… and never considered
Linux clients.)

That said, the absence of the CSD stub *doesn't even matter* for those who
are running a recent version of `csd-post.sh`, which entirely sidesteps
running the server provided stub and binaries.

Long story short: This patch makes OpenConnect not fail if the CSD stub
can't be downloaded… as long as a CSD wrapper script was specified.

Signed-off-by: Daniel Lenski <dlenski@gmail.com>