Make csd-post.sh cope with not being able to write to stdout
When invoked from gnome-shell for NetworkManager, curl takes a SIGPIPE
when writing the result of the final POST to stdout. It then exits with
result code 23, causing OpenConnect to whine about an error.

Make it return zero.

Also stop the pidof results making it to stdout when checking for a
process.

Signed-off-by: David Woodhouse <dwmw2@infradead.org>