defer the switch to syslog until AFTER the tunnel is fully up,
This way, initial connection information and background PID will be usefully
printed to the console, as will errors which prevent the tunnel from being
started (and thus cause OpenConnect to abort as soon as it's established
a connection to the server).

This is a further refinement of 557ac6cfa6d42045ac5ed5e65e2fb079106643fa (!117).

Signed-off-by: Daniel Lenski <dlenski@gmail.com>